feat: remove legacy external
