+++ src/runtime/applicationinsights.ts
@@ -1,0 +1,4 @@
+import * as applicationinsights from 'applicationinsights';
+export * from "applicationinsights"
+// @ts-ignore wtf applicationinsights ????
+export default applicationinsights.default
