--- src/runtime/setup.ts
@@ -1,51 +1,0 @@
-import * as applicationInsights from 'applicationinsights'
-import type { TNitroAppInsightsConfig } from '../types'
-
-export function setup (config: TNitroAppInsightsConfig) {
-  // Setup Application Insights using the instrumentation key from the environment variables
-  const configuration = applicationInsights
-    .setup(config.connectionString)
-    .setAutoCollectRequests(config.autoCollectRequests)
-    .setAutoCollectDependencies(config.autoCollectDependencies)
-    .setAutoCollectExceptions(config.autoCollectExceptions)
-    .setAutoCollectHeartbeat(config.autoCollectHeartbeat)
-    .setAutoCollectIncomingRequestAzureFunctions(config.autoCollectIncomingRequestAzureFunctions)
-    .setAutoCollectPreAggregatedMetrics(config.autoCollectPreAggregatedMetrics)
-    .setDistributedTracingMode(config.distributedTracingMode)
-    .setSendLiveMetrics(config.sendLiveMetrics)
-    .setUseDiskRetryCaching(config.useDiskRetryCaching)
-
-  if (typeof config.autoCollectPerformance === 'object') {
-    configuration.setAutoCollectPerformance(
-      config.autoCollectPerformance.value, config.autoCollectPerformance.collectExtendedMetrics)
-  } else {
-    configuration.setAutoCollectPerformance(
-      config.autoCollectPerformance)
-  }
-
-  if (typeof config.autoDependencyCorrelation === 'object') {
-    configuration.setAutoDependencyCorrelation(config.autoDependencyCorrelation.value, config.autoDependencyCorrelation.useAsyncHooks)
-  } else {
-    configuration.setAutoDependencyCorrelation(config.autoDependencyCorrelation)
-  }
-
-  if (typeof config.enableWebInstrumentation === 'object') {
-    configuration.enableWebInstrumentation(config.enableWebInstrumentation.value, config.enableWebInstrumentation.WebSnippetConnectionString)
-  } else {
-    configuration.enableWebInstrumentation(config.enableWebInstrumentation)
-  }
-
-  if (typeof config.autoCollectConsole === 'object') {
-    configuration.setAutoCollectConsole(config.autoCollectConsole.value, config.autoCollectConsole.collectConsoleLogs)
-  } else {
-    configuration.setAutoCollectConsole(config.autoCollectConsole)
-  }
-
-  if (typeof config.internalLogging === 'object') {
-    configuration.setInternalLogging(config.internalLogging.enableDebugLogging, config.internalLogging.enableWarningLogging)
-  } else {
-    configuration.setInternalLogging(config.internalLogging)
-  }
-
-  return configuration.start()
-}
